Add files via upload
This is the main program where I use the API. Note that the API only allows 5 requests a minute so I can either only run it once and get top 5 losers or I can create another loop so that in 5 minutes I get the top 25 losers. I have yet to do the latter. Another thing to notice is that the Trading View API only extracts data for US stocks data.