MainProgram.py
This is the main program using the Trading View API that helps find the income change in the most recent quarter of the top 5 losers. I would do more but it gets complicated since it only does US stocks and many of the losers can be on other markets. Forgive me as I did not optimize my code and I did not clean it up :)